+++ methods/src/main/kotlin/methods/inequalities/InequalitiesRules.kt
@@ -4,13 +4,16 @@ import engine.conditions.Sign
 import engine.conditions.signOf
 import engine.expressions.Constants
 import engine.expressions.Fraction
+import engine.expressions.Variable
+import engine.expressions.contradictionOf
 import engine.expressions.fractionOf
+import engine.expressions.identityOf
 import engine.expressions.inverse
 import engine.expressions.isNeg
 import engine.expressions.productOf
+import engine.expressions.setSolutionOf
 import engine.expressions.simplifiedNegOf
-import engine.expressions.solutionOf
-import engine.expressions.xp
+import engine.expressions.variableListOf
 import engine.methods.Rule
 import engine.methods.RunnerMethod
 import engine.methods.rule
@@ -35,12 +38,12 @@ enum class InequalitiesRules(override val runner: Rule) : RunnerMethod {
             onPattern(inequality) {
                 if (inequality.holdsFor(getValue(lhs), getValue(rhs))) {
                     ruleResult(
-                        toExpr = solutionOf(xp(context.solutionVariables.first()), Constants.Reals),
+                        toExpr = identityOf(variableListOf(context.solutionVariables), get(inequality)),
                         explanation = metadata(Explanation.ExtractSolutionFromTrueInequality),
                     )
                 } else {
                     ruleResult(
-                        toExpr = solutionOf(xp(context.solutionVariables.first()), Constants.EmptySet),
+                        toExpr = contradictionOf(variableListOf(context.solutionVariables), get(inequality)),
                         explanation = metadata(Explanation.ExtractSolutionFromFalseInequality),
                     )
                 }
@@ -62,12 +65,12 @@ enum class InequalitiesRules(override val runner: Rule) : RunnerMethod {
                 if (lhsSign.isKnown() && rhsSign.isKnown()) {
                     if (inequality.holdsFor(lhsSign.signum.toBigDecimal(), rhsSign.signum.toBigDecimal())) {
                         ruleResult(
-                            toExpr = solutionOf(xp(context.solutionVariables.first()), Constants.Reals),
+                            toExpr = identityOf(variableListOf(context.solutionVariables), get(inequality)),
                             explanation = metadata(Explanation.ExtractSolutionFromTrueInequality),
                         )
                     } else {
                         ruleResult(
-                            toExpr = solutionOf(xp(context.solutionVariables.first()), Constants.EmptySet),
+                            toExpr = contradictionOf(variableListOf(context.solutionVariables), get(inequality)),
                             explanation = metadata(Explanation.ExtractSolutionFromFalseInequality),
                         )
                     }
@@ -87,7 +90,7 @@ enum class InequalitiesRules(override val runner: Rule) : RunnerMethod {
 
             onPattern(inequality) {
                 ruleResult(
-                    toExpr = solutionOf(move(lhs), inequality.toInterval(get(rhs))),
+                    toExpr = setSolutionOf(variableListOf(move(lhs) as Variable), inequality.toInterval(get(rhs))),
                     explanation = metadata(Explanation.ExtractSolutionFromInequalityInSolvedForm),
                 )
             }